Feature: Book Card folder change & Styled
